requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.21
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
fake-useragent==2.2.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==58.1.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
zipp==3.22.0
